requirements.txt
aiogram==3.15.0
aiohttp==3.10.11
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
redis==5.2.1
asyncpg==0.30.0